파이널 전 최정현

+++ src/main/java/com/example/project_team/dto/AccountDTO.java
@@ -23,7 +23,7 @@ public class AccountDTO {
 	private long accountNum;         //계좌번호 pk
 	private String id;				//회원id
 	private int accountPW;          //계좌 비밀번호
-	private int balance;            //잔액
+	private long balance;            //잔액
 	private char accountType;       //계좌종류(디폴트 입출금:e, 예금: y , 적금: j, 대출: d, 펀드: p)
 	private char accountState;      //계좌상태(정상 : j, 휴먼 : h, 해지 : c, 정지 : s)
 	private int accountLimit;       //일일한도
